Add CVS issues and script to checkout revisions in reverse order
A CVS versions file got corrupted long ago, but it was still usable
and accumulated more revisions. The problem become visible in the
migration process (see CVS-issues.txt). It was necessary to checkout
the revisions in reverse order (see co_cvs_revs) to find
the problematic one, then understand the problem in the diffs and fix
the version file in order to be able to checkout all revisions and
complete the migration.