Make possible to create and switch to branch
For the . branch to work for submodules, both the superproject and the
submodule should have the same branch, so the branch also needs exist in
the superproject. The new operations branch and branch_sw allow
creation of branch and switch to branch respectively, so that before
revisions for a new Slackware release are created the branch for the old
version is created and the branch for the new release is created and
switched to.

See https://git-scm.com/docs/git-submodule#Documentation/git-submodule.txt--bltbranchgt