Rewrite revisions, create and switch branches
Rewrite original revisions on the main branch, then create SLACK-10_2
and switch to SLACK-11_0 branch. This allows the superproject repository
to be checked out for specific Slackware release. For example:

  git clone https://github.com/gdsotirov/slackbuilds.git
  cd slackbuilds
  git submodule init
  git submodule update
  git switch SLACK-11_0
  git submodule update --remote
  git switch SLACK-10_2
  git submodule update --remote
  git switch main
  git submodule update --remote

Remark: After moving from newer to older branch (e.g. from SLACK-11_0 to
SLACK-10_2) submodule directories that do not exist in older branch
would have to be removed manually.

The above is necessary, because the superproject repository would be
checked out in different branches on different build machines with
different Slackware versions like before with CVS.